delete all the css files from the pages and components I worked on, to replace them with styled components exported from dedicated files.
