+++ src/components/App/App.component.jsx
@@ -1,7 +1,7 @@
 import React from 'react';
 import { BrowserRouter, Switch, Route } from 'react-router-dom';
 import { ThemeProvider } from 'styled-components';
-
+import GlobalStyle from './styled';
 import AuthProvider from '../../providers/Auth';
 import HomePage from '../../pages/Home';
 import LoginPage from '../../pages/Login';
@@ -9,18 +9,14 @@ import NotFound from '../../pages/NotFound';
 import SecretPage from '../../pages/Secret';
 import Private from '../Private';
 import Layout from '../Layout';
-
-const theme = {
-  main: '#1c5476',
-  mainFade: '#2e729d',
-  secondary: 'white',
-};
+import theme from '../../theme';
 
 function App() {
   return (
     <BrowserRouter>
       <AuthProvider>
         <ThemeProvider theme={theme}>
+          <GlobalStyle />
           <Layout>
             <Switch>
               <Route exact path="/">
