+++ src/components/Avatar/styled.js
@@ -1,0 +1,7 @@
+import styled from 'styled-components';
+
+export const Image = styled.img`
+  width: 40px;
+  height: 40px;
+  border-radius: 50%;
+`;
