+++ src/components/Content/Content.component.jsx
@@ -1,21 +1,14 @@
 import React from 'react';
-import styled from 'styled-components';
+import { Container } from './styled';
 import VideoCard from '../VideoCard';
-
-const StyledContent = styled.div`
-  display: flex;
-  flex-wrap: wrap;
-  justify-content: center;
-  align-items: center;
-`;
 
 const Content = ({ items }) => {
   return (
-    <StyledContent>
+    <Container>
       {items.map((item) => (
         <VideoCard key={item.etag} data={item} />
       ))}
-    </StyledContent>
+    </Container>
   );
 };
