+++ src/components/Icons/SunIcon/Sun.icon.jsx
@@ -1,14 +1,5 @@
 import React from 'react';
-import styled from 'styled-components';
-
-const StyledSun = styled.svg.attrs((props) => ({
-  xmlns: 'http://www.w3.org/2000/svg',
-  enableBackground: 'new 0 0 24 24',
-  viewBox: '0 0 24 24',
-  fill: props.theme.secondary,
-  width: `${props.size}px`,
-  height: `${props.size}px`,
-}))``;
+import StyledSun from './styled';
 
 const Sun = ({ size = 25 }) => {
   return (
