--- src/components/Layout/Layout.styles.css
@@ -1,4 +1,0 @@
-.container {
-  width: 100vw;
-  height: 100vh;
-}
