+++ src/components/Layout/styled.js
@@ -1,0 +1,6 @@
+import styled from 'styled-components';
+
+export const Container = styled.main`
+  width: 100vw;
+  height: 100vh;
+`;
