+++ src/index.jsx
@@ -2,7 +2,6 @@ import React from 'react';
 import ReactDOM from 'react-dom';
 
 import App from './components/App';
-import './global.css';
 
 ReactDOM.render(
   <React.StrictMode>
