+++ src/theme/index.js
@@ -1,0 +1,7 @@
+const theme = {
+  main: '#1c5476',
+  mainFade: '#2e729d',
+  secondary: 'white',
+};
+
+export default theme;
